feat(variables): implement process_variable_value function and refactor variable handling
